requirements.txt
# Core UI Framework
PySide6>=6.6.0

# HTTP Client for Ollama (httpx supports both sync and async)
httpx>=0.27.0

# Process management
psutil>=5.9.0

# RAG (Retrieval Augmented Generation) - Built-in Memory System
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Optional: Backend API (for future extensions)
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-dotenv>=1.0.0
PyYAML>=5.1
